Kubelet: only sync iptables on linux
